remove priority queue
